feat: created seperate files for skeletons
