+++ src/components/skeleton/index.ts
@@ -1,21 +1,15 @@
 // components/skeleton/index.ts
 
-// Since all components are in one file, import them all from that file
-// Assuming your large file is named 'SkeletonComponents.tsx'
-export {
-  Skeleton,
-  SkeletonText,
-  SkeletonCard,
-  SkeletonAvatar,
-  SkeletonButton,
-  SkeletonBreadcrumb,
-  SkeletonHeader,
-  SinglePostSkeleton,
-  NewsletterArchiveSkeleton,
-  NewsletterViewSkeleton,
-  LeadsViewSkeleton,
-  DashboardViewSkeleton,
-  KanbanBoardSkeleton,
-} from './SkeletonBase.tsx';
-
-// OR if you named the file differently, adjust the import path accordingly
+export { Skeleton } from './SkeletonBase';
+export { SkeletonText } from './SkeletonText';
+export { SkeletonCard } from './SkeletonCard';
+export { SkeletonAvatar } from './SkeletonAvatar';
+export { SkeletonButton } from './SkeletonButton';
+export { SkeletonBreadcrumb } from './SkeletonBreadcrumb';
+export { SkeletonHeader } from './SkeletonHeader';
+export { SinglePostSkeleton } from './SinglePostSkeleton';
+export { NewsletterArchiveSkeleton } from './NewsletterArchiveSkeleton';
+export { NewsletterViewSkeleton } from './NewsletterViewSkeleton';
+export { LeadsViewSkeleton } from './LeadsViewSkeleton';
+export { DashboardViewSkeleton } from './DashboardViewSkeleton';
+export { KanbanBoardSkeleton } from './KanbanBoardSkeleton';
